Heart_Attack_project/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
ansi2html==1.8.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.1.0
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasist==1.5.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dicttoxml==1.7.16
dill==0.3.8
docopt==0.6.2
docutils==0.20.1
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.2.0
fastjsonschema==2.19.0
Flask==3.0.0
fonttools==4.45.0
fqdn==1.5.1
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geoip2==4.8.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
idna==3.4
imbalanced-learn==0.12.0
imblearn==0.0
importlib-metadata==7.0.0
ip2geotools==0.1.6
IP2Location==8.10.2
ipykernel==6.25.0
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.1.2
jaraco.classes==3.3.1
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==24.3.1
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
maxminddb==2.5.2
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
module-name==0.6.0
more-itertools==10.2.0
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.5
mysql-connector-python==8.2.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.7
nh3==0.2.15
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
openpyxl==3.1.2
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.1
pandas==2.1.2
pandocfilters==1.5.0
parse==1.19.1
parso==0.8.3
patsy==0.5.6
pickleshare==0.7.5
Pillow==10.1.0
pip-review==1.3.0
pipreqs==0.5.0
pkginfo==1.9.6
platformdirs==3.10.0
plotly==5.18.0
prometheus-client==0.19.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
pyarrow==15.0.0
PyAutoGUI==0.9.54
pycodestyle==2.11.1
pycountry==23.12.11
pycparser==2.21
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.15.1
pylint==3.1.0
PyMsgBox==1.0.9
pyodbc==5.1.0
pyparsing==3.1.1
pyperclip==1.8.2
pyquery==2.0.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.5.1
QtPy==2.4.1
ratelim==0.1.6
readme_renderer==43.0
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
selenium==4.18.1
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.2
statsmodels==0.14.1
streamlit==1.31.0
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
toolz==0.12.1
tornado==6.3.2
tqdm==4.66.2
traitlets==5.9.0
trio==0.24.0
trio-websocket==0.11.1
twine==5.0.0
typed-ast==1.5.5
types-python-dateutil==2.8.19.14
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ua-parser==0.18.0
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.0.7
user-agents==2.2.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
